first revision of the code; folder and path anonymization
