--- ASR/labels/ASC/dev.txt
@@ -1,100 +1,0 @@
-بل يجب لبحث فيما تمثله من إضافةن حقيقيةن للإقتصاد لمصري  في مجالات لتوظيف  باعتبار أن مشكلة لبطالة من لمشكلات لرئيسية في مصر
-واختتم لبرازيليون مباراياتهم لإعدادية بالفوز على صربيا بهدفن وحيدن  سجله لمهاجم فريدن في لشوط لثاني من لمباراة  ألتي أقيمت في ساوباولو
-وأضاف لبيان ن فريقن من لأطباء والممرضات  وإختصاصيين أاخرين في مجال لصحة  يعتنون بمانديلا على مدار لساعة
-تربية لحمام  هوايةن ومهنةن لبعض لناس
-لم يمنع ضعف موارد لتمويل  وارتفاع كلفة لمشاركات لدولية  رياضة لفروسية في تونس من أن تستقطب لمئات  من عشاقها في بلدن يكاد لإهتمام فيه يقتصر  على رياضاتن شعبيةن معينةن
-وتطور ستخدام لطائرات لعاملة بدون طيار  وبدأت لساعات لذكية لإنتشار  وكذلك لطباعة لثلاثية لأبعاد
-وفي كوريا لجنوبية تقوم لحكومة بتمويل مستشفياتن لعلاج هذا لإدمان لذي يعتبر مشكلةن قوميةن
-هناك فكرةن طرحت بادئ لأمر  بعقد قمةن أوروبيةن في سرييفو بهذه لمناسبة
-وكتبوا في دورية لكلية لأميركية لأمراض لقلب  أن لسمنة ترتبط بحدوث تغييرات  في لقلب لدى لبالغين
-شهد عام ألفين وأربعة عشرة  عدة إنجازاتن طبيةن
-عاد إرتفاع أسعار لأدوية  وشح لمنقذ للحياة منها  ليطل برأسه في لسودان من جديدن
-ويقول لعلماء إنه من غير لمرجح أن تطور لبكتيريا لمعدية  مقاومةن ضد لعلاج لجديد  ألذي أصبح متاحن بالفعل  في شكل مرهمن للأمراض لجلدية
-لقد ساهمت لتكنلوجيا في تقليل لنزاعات لأسرية  وأعطت لكل فردن نوعن من لإستقلالية
-ويستخدم لبرنامج  نظامن سحابين للذكاء لصنعي  يسمح له بتحليل لإيماءات  والتعابير
-كما شاركت طالباتن من مدارس فلسطينيةن  ألفنانة لتركية في عمل لوحات
-في حين كان لمركز لأول في لوثب لعالي  من نصيب لكرواتية أانا سيميتش
-وبالتالي تساعد على لوقاية من لإمساك
-أكدت عشر دولن عربيةن شاركت لخميس لماضي في جتماع جدة  موافقتها على لإنضمام إلى حلفن مع لولايات لمتحدة  لشن حملةن عسكريةن منسقةن ضد تنظيم  ألدولة لإسلامية
-وبات أتلتيكو مدريد في صدارة لترتيب لعام  بأربع نقاط
-أتاحت للبائع لمتجول أن يكون جاذبن للمواطن لأقل دخلن
-أعلنت وزارة لصحة لسعودية ليوم عن وفاتين جديدتين بالفيروس لتاجي  كورونا نوفل
-هل للصيام تأثيرن على عمل لمسلمين في لشركات بأوروبا
-وكل ذلك بمحتوىن منخفضن للغاية من لسعرات لحرارية
-ويمكنك لحصول على تطبيقاتن للتدريبات لأساسية مجانن
-من بينها ألإستقرار ونوعية لرعاية لصحية والثقافة والبيئة والتعليم والبنية لتحتية
-وكان لأمل أن تكون هذه لديمقراطيات  مصحوبةن بأداءن تنموين مختلف
-ليضعا بذلك حدن  للعديد من لتقارير  ألتي أكدت إمكانية رحيل للاعب لمشاغب قريبن
-وتستخدم في بعض لدول  وسائل علاجن مختلفةن
-وأصبح هذا لصرح متحفن رسمين
-تحتضن قاعة  ذا فينيو وسط بيروت  معرض لفن لإستثنائي
-واشتهر لراحل بمقالاته وكتبه لرصينة  ألتي تضمنت قراءاتن مستقبليةن للأافاق لسياسية والإجتماعية  في لعالم لعربي لإسلامي
-ألذي يتضمن مزايا برمجيةن وبصريةن عديدةن  تهدف لمواكبة لتطور لحاصل في لفضاء لإلكتروني  وتسهيل ستفادة لقراء من خدمات لموقع
-وذلك في بطولة كأس لعالم للأندية بعد نسخةن تاريخيةن ألعام لماضي  إنتهت بتتويج بايرن ميونيخ لألماني  على حساب لرجاء لمغربي  في أول تأهلن لفريقن عربين إلى نهائي لمسابقة
-إفتتحت لجمعة فعاليات لدورة لرابعة عشرة  من لمهرجان لدولي للفيلم بمراكش
-وذلك بزيارة جمهورن خاصن جدن سنوين
-إلتحق لوكا زيدان  إبن لنجم لدولي لفرنسي لجزائري لأصل زين لدين زيدان  بالفريق
-حدث ذلك في حي  يعقوب لمنصور لشعبي
-ويأمل لباحثون  تطوير حبوبن أو نسخةن من لدواء  قابلةن للحقن خلال خمس سنوات
-لامس مذنبن يطلق عليه إسم  سايدينغ سبرينغ  كوكب لمريخ عند مروره بمحاذاته
-على عتبارها تساعد على توسيع مدارك لأطفال  وتجعل منهم  أناسن مثقفين مستقبلن  ومواكبين لعصر تكنولوجيا لمعلومات
-ويعتبر مهرجان قرطاج لسينمائي  من أعرق مهرجانات أفريقيا
-لكن مصدرن في لوفد قال  إن لسعر سينخفض بعد نخفاض أسعار لنفط في لعالم
-شرع لإتحاد لتونسي للفروسية في تنفيذ  خطةن ترنو إلى لمضي بهذه لرياضة  نحو بلوغ لعالمية
-أخفق مجلس لنواب للبناني في ختيار رئيسن جديدن للبلاد  خلفن للرئيس لحالي لذي تنتهي ولايته في لخامس والعشرين  من مايو أيار لمقبل
-وبعد شهورن من لحيرة والقلق  تعرف كومار على مركز عبد لله بن زيد لثقافي  للتعريف بالإسلام
-إذ سيحضر لقاء هذا لعام خمسن وثلاثون منهم
-وبسبب شكوكن بأن لطائرة كانت تقل دوارد سنودن  ألذي تتهمه واشنطن بالتجسس
-وأوضح لباحثون أن سوء لتغذية  هو لسبب لرئيسي لتوقف لنمو  عند لأطفال
-فتحظى بتعلمن يمحو أميتها  ويضيء لها طريق لمعرفة والتكنولوجيا
-ومنح أابائهم لقدرة على لتحكم في كيفية ستخدام هذه لخدمة
-كما تم تحسين واجهات لتنقل  واختيار وسائل لنقل لمناسبة بشكلن كبير
-ولعبت لخبرة دورها في تتويج عاشور لخامس عالمين
-قبل أن يسلم ليدافع عن دينه  محبن ومحترمن لأصله وماضيه
-شارك موقع  ألجزيرة لتعلم لعربية  في لمؤتمر لدولي لثاني للغة لعربية  ألذي نظمته جامعة مولانا مالك إبراهيم  ألإسلامية لحكومية بمدينة مالانق  بالتعاون مع جامعة دار لسلام  بمدينة كونتور  في أندونيسيا
-وذلك على خلاف نظرائه لسابقين
-دون تحميل لجسم جهدن كبيرن في لبداية  قد يتسبب في نفور لشخص من لإستمرار
-تسعى شبكة لتواصل لإجتماعي لصاعدة  إيلو  إلى منافسة شبكة  فيسبوك  عبر لتخلي عن لإعلانات  والحفاظ على لخصوصية  وحماية لبيانات
-مضيفن  نحاول خلق فرص عملن بأيدينا
-رفض مجلس لأمن لدولي  مشروع لقرار لفلسطيني لرامي  إلى إنهاء لإحتلال لإسرائيلي في عامين
-مما يدعم نتائج دراساتن سابقةن  تحذر من مخاطر لإفراط في ستعمال لجوال
-وصدرت لرسائل  بإسم مبدعي وفناني مصر
-و يمكن أن تحصد لثمار  على مدى فترةن زمنيةن طويلةن
-ألتوحد هو مجموعة ضطراباتن عصبيةن في لتطور  تشمل أعراضها وجود مشاكل في لسلوك لاجتماعي  للشخص لمصاب
-والميكونغ وسألوين في جنوب شرق أاسيا
-وذلك بحضور رئيس لهيئة
-ألفكرة وإن بدت جديدةن على مجتمعن يعيش واقعن سيئن  لا تشجع على لضحك
-فتحت أكاديمية لموسيقى لعربية رسمين يوم لسبت  أبوابها في بروكسل  بحضور مجموعةن من لوزراء  ورجال لفن لبلجيكيين والعرب
-رمز لدولة لتركية لعلمانية لتي تأسست  عقب نهيار لدولة لعثمانية
-تتوالى لعمليات لسرية بالحدوث
-قال باحثون إن رياحن أقوى من لمعتاد  خففت من حرارة سطح لمحيط لهادي  هي سبب لتباطئ لمؤقت  في رتفاع درجة حرارة لأرض  منذ بداية لقرن لحادي والعشرين
-فهذا لمنزل لمتواضع  أصبح محجن لعددن كبيرن من لنساء لمريضات بالسرطان
-ومن تلك لسلع  ألشاي لصيني  والورق والبارود والبوصلة
-كما توفيت لروائية لبارزة  والأستاذة لجامعية لمصرية  رضوى عاشور  عن ثماني وستين عامن
-وأن أهم معوقات نجاحها تكمن في عدم تفرغ أصحابها لإدارتها
-وبعثوا رسالةن إلى لرئيس تتضمن مطالب لعودتهم
-وبهذا يبقى مئةن وستةن وثلأثون محتجزن في لمعتقل لمثير للجدل
-أعلنت لحكومة لمصرية عن وقف تقديم لدعم لنقدي لمزارعي لقطن  إعتبارن من لموسم لزراعي لمقبل
-أحرزت منتخبات لبرازيل وألمانيا وروسيا  فوزن في مقابلاتهم لإعدادية لتي أقيمت ستعدادن لنهائيات كأس لعالم  ألتي ستنطلق بعد أقل من أسبوع
-إنه عندما يكون أحد لزوجين يستخدم أحد أشكال لتكنولوجيا  أكثر من لأاخر
-واعتبروها فلسفةن روحيةن متكاملةن لتحرير لجسم والفكر
-وأحرز لمركز لثالث  ألروائي لجزائري  أحمد طيباوي عن روايته  موتن ناعم
-رغم لنزاع لدامي  ألذي يعصف بالبلاد  منذ ثلأث سنوات
-وفي فتتاح لمؤتمر  قالت لشاعرة شريفة لسيد  إن لعقاد تخذ من لقراءة والإطلاع على كل لعلوم  ومختلف لحضارات  سلاحن يحطم به لصنمية  ويكسر لمحرمات
-وذلك مقارنةن مع لمحاصيل لزراعية لأخرى
-وجاء هذا لقرار  بعد إعلان لسعودية تخفيض أعداد لحجاج هذا لعام
-جمع لمؤنث لسالم مثل  فازت إحدى لطالبات في مسابقة لقراءات لقرأانية
-دخلت لتكنولوجيا في كل بيتن وأسرةن  وأصبحت تشكل لجزء لكبير من حياتنا
-ألمشاكل لتي يتركها خلفه دائمن
-منها أقمشةن وأدواتن معدنيةن وخشبيةن  وقنانن بلاستيكيةن وزجاجيةن وأوراق صحف
-حيث كان لعرب أول من عرف قيمتها لعلاجية  في لقرن لعاشر قبل لميلاد  في مملكة سبأ
-حتى لأمس لقريب  كان لكثير من لأوكرانيين  يشككون في نتماء تتار شبه جزيرة لقرم
-وكلما زادت كمية لسكر لمتناولة مع لتمر تقل فائدته لغذائية
-حذرت لأمم لمتحدة من أن لعالم سيواجه خلال لعقود لمقبلة  تفاقم أزمةن مزدوجةن في لمياه  والكهرباء
-حضور بعض لزعماء في  ألمسيرة لجمهورية  بباريس
-وأوضحت لمجلة أن لسبب في ذلك  يرجع إلى تضيق لشعب لهوائية وتشنجها بفعل لهواء لبارد
-ولا يزال هذا لبلد متمسكن بتقويم لكنيسة لقبطية  ألمعروف محلين بالتقويم لإثيوبي
-وجاءت لأرقام صادمةن  في ما يخص لشرق لأوسط
-مشيرن إلى خدمة لقرأان لكريم  وتعزيز علاقة لمسلمين به
-والعمل لرئيسي له هو روايته لملحمية  مائة عامن من لعزلة  ألتي نال عنها جائزة نوبل في لأدب  عام ألفن وتسعمئةن وإثنان وثمانون
-ملقين لضوء على قضية لفتنة لطائفية  في لمجتمع لمصري بأسلوبن بسيطن  من خلال علاقات لأطفال  في لمدرسة بزميلهم لمسيحي
