--- TTS/hubert_labels/ASC/valid.tsv
@@ -1,101 +1,0 @@
-/l/users/speech_lab/ArabicSpeechCorpus/ArabicSpeechCorpus16K/		
-test/ARA NORM  0001.wav	113730	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0002.wav	284055	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0003.wav	258370	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0004.wav	77975	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0005.wav	181911	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0006.wav	169965	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0007.wav	160407	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0008.wav	345836	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0009.wav	166551	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0010.wav	63639	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0011.wav	256322	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0012.wav	142743	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0013.wav	110658	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0014.wav	140012	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0015.wav	49303	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0016.wav	358722	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0017.wav	271682	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0018.wav	137965	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0019.wav	89154	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0020.wav	303682	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0021.wav	144279	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0022.wav	155202	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0023.wav	202903	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0024.wav	112706	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0025.wav	172525	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0026.wav	137623	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0027.wav	139671	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0028.wav	117314	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0029.wav	190530	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0030.wav	146754	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0031.wav	458818	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0032.wav	202562	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0033.wav	111426	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0034.wav	173549	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0035.wav	234220	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0036.wav	60909	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0037.wav	246338	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0038.wav	134551	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0039.wav	138818	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0040.wav	88983	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0041.wav	131820	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0042.wav	304109	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0043.wav	122007	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0044.wav	141122	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0045.wav	211607	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0046.wav	143511	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0047.wav	158956	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0048.wav	172012	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0049.wav	142060	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0050.wav	360940	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0051.wav	172098	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0052.wav	77378	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0053.wav	81303	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0054.wav	257175	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0055.wav	142572	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0056.wav	161431	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0057.wav	177333	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0058.wav	95725	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0059.wav	142658	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0060.wav	104429	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0061.wav	172695	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0062.wav	276972	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0063.wav	265794	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0064.wav	57922	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0065.wav	226029	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0066.wav	98455	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0067.wav	222530	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0068.wav	280556	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0069.wav	70039	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0070.wav	124567	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0071.wav	156603	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0072.wav	183874	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0073.wav	110743	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0074.wav	69186	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0075.wav	80450	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0076.wav	157335	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0077.wav	91116	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0078.wav	200599	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0079.wav	124482	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0080.wav	92226	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0081.wav	218605	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0082.wav	127554	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0083.wav	94956	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0084.wav	85655	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0085.wav	330562	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0086.wav	174658	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0087.wav	138306	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0088.wav	204183	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0089.wav	101100	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0090.wav	131479	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0091.wav	166807	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0092.wav	101356	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0093.wav	63469	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0094.wav	114498	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0095.wav	114925	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0096.wav	162882	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0097.wav	197783	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0098.wav	124652	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0099.wav	291223	speaker_embedding/ASC_speaker_embedding.npy
-test/ARA NORM  0100.wav	114241	speaker_embedding/ASC_speaker_embedding.npy
