--- dict.txt
@@ -1,92 +1,0 @@
-▁ 7888394
-ا 5240837
-ل 4086510
-ي 2961895
-م 2193275
-ن 2167596
-و 1852812
-ت 1552697
-ر 1543221
-ع 1224967
-ب 1116473
-ه 1088784
-د 1060960
-ة 1037790
-أ 940870
-س 935180
-ف 834969
-ك 831374
-ق 807989
-ح 704969
-ج 436806
-ذ 368254
-ط 344915
-إ 327964
-ش 314812
-ى 307987
-ص 290514
-خ 277022
-ض 252120
-ث 197381
-ز 166103
-ئ 137402
-ً 122431
-غ 113416
-ء 111245
-ظ 98346
-ُ 75189
-آ 58959
-ؤ 57561
-ّ 39925
-0 27260
-ٍ 24728
-َ 22922
-ِ 21587
-1 18933
-2 15936
-ٌ 10327
-5 6938
-9 6502
-6 4829
-7 4380
-8 4041
-e 3578
-a 3330
-% 2892
-ـ 2768
-t 2746
-i 2683
-n 2462
-o 2390
-r 2341
-s 2073
-c 1561
-l 1504
-m 1141
-d 1018
-p 988
-u 920
-h 888
-g 777
-b 762
-f 603
-y 562
-k 515
-w 470
-v 293
-j 254
-z 193
-x 105
-@ 97
-3 52
-4 48
-q 30
-̇ 6
-ٱ 2
-⁄ 1
-madeupword0000 0
-madeupword0001 0
-madeupword0002 0
-madeupword0003 0
-madeupword0004 0
-madeupword0005 0
